basic-agent-tracing-langsmith/pyproject.toml
[project]
name = "basic-agent-tracing-langsmith"
version = "0.1.0"
description = ""
authors = [
    {name = "Sumit Ghosh"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.20,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "langsmith (>=0.4.2,<0.5.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
